[clangd] Extend clangd's hover information with allocation info
This commit adds two new clangd compile parameters:
1. CLANGD_ALLOCATION_BUCKETS specifies the memory allocation buckets
2. CLANGD_ALLOCATION_TRACKING_BYTES specifies the tracking overhead
   for every allocation.
Both are combined to calculate the raw allocation size and the resulting
allocation bucket. Finally this information is used to calculate
the memory utilization, which basically represents internal
fragmentation for every allocation.

This information is now displayed when hovering over a type.