[clangd] Add size estimation for templated structs
clangd completely bails our to estimate the size
and the alignment for structs that somehow depend
on a template.

With this commit, we walk a struct's member variables
to calculate alignment and size on the fly.

While the real fix should address the initial parsing.
This change is to the very least a quality of life
improvement.