[clangd] Add background indexing after ClangdLSPServer::onInitialize()
When starting clangd, it does not seem to index all known files.
Consequently not all relations and symbols are known immediately.
This leads to incomplete results when, for instance, listing a
type's hierarchy.

For small code bases, we can easily index all files on startup
without taking a significant hit in memory consumption.

This commit extends ClangdLSPServer::onInitialize() to parse
compile_commands.json and then schedules all files stated in
the DB for background indexing.
This feature requires --compile-commands-dir to be set.
Otherwise, no files will be schedules for indexing.