[clangd] Fix type hierarchy for templated derivation
Considering the following example
    template <typename T> struct foo {};
    struct bar : foo<int>{};

type hierarchy for bar is broken as it finds that
it derives from foo<int>. Template specializations
are however not stored in the relations.

getTypeHierarchy() faces the same problem and
drops the specialization already. fillSuperTypes()
is missing this step. After adding it there as well,
the type hierarchy works.